fix(discord-voice): route status→work, extract voice tools, fix mode-indicator
- open_github_url narrowed to OPEN-only; PR/issue STATUS questions route to the
  work tool instead of opening a tab every time
- get_task_status removed (it was exempt from the dispatch gate and fired
  spontaneously, narrating "still working in the background")
- send_discord_message added (owner-tier, dispatch-gated, allowlist channels) —
  posts directly via the voice bot's discord.js client
- extracted open_github_url / switch_mode / dismiss / share_screen /
  send_discord_message into discord-voice-tools.ts (voice-session-scoped tools
  that need `s`; server consts passed via options to avoid a circular import)
- mode indicator stays on the nickname; retry on a failed/throttled nickname
  update (was stuck), and strip ALL stacked mode-emojis (+ variation selectors)
  to fix the 🔇🗣 stacking
